fix(web): split story helper data exports
